test(management): share route test monque fixture
