+++ packages/management/tests/unit/orpc-capabilities.test.ts
@@ -1,45 +1,8 @@
 import { describe, expect, test } from 'vitest';
 
+import { createManagementMonque } from '@tests/unit/management-test-utils';
 import { createManagementSurface } from '@/index';
-import type { ManagementMonque, ManagementOpenApiContext, ManagementSurface } from '@/surface';
-
-function createManagementMonque(
-	overrides: Partial<ManagementMonque> = {},
-	options: { mutations?: boolean } = { mutations: true },
-): ManagementMonque {
-	const monque: ManagementMonque = {
-		isHealthy: () => true,
-		getQueueViewSummaries: async () => [],
-		getJobsWithCursor: async () => ({
-			jobs: [],
-			cursor: null,
-			hasNextPage: false,
-			hasPreviousPage: false,
-		}),
-		getJob: async () => null,
-		getQueueStats: async () => ({
-			pending: 0,
-			processing: 0,
-			completed: 0,
-			failed: 0,
-			cancelled: 0,
-			total: 0,
-		}),
-		...overrides,
-	};
-
-	if (options.mutations ?? true) {
-		monque.cancelJob = async () => null;
-		monque.retryJob = async () => null;
-		monque.rescheduleJob = async () => null;
-		monque.deleteJob = async () => false;
-		monque.cancelJobs = async () => ({ count: 0, errors: [] });
-		monque.retryJobs = async () => ({ count: 0, errors: [] });
-		monque.deleteJobs = async () => ({ count: 0, errors: [] });
-	}
-
-	return monque;
-}
+import type { ManagementOpenApiContext, ManagementSurface } from '@/surface';
 
 async function handleCapabilities(surface: ManagementSurface, context?: ManagementOpenApiContext) {
 	const result = await surface.openApiHandler.handle(
@@ -57,7 +20,7 @@ async function handleCapabilities(surface: ManagementSurface, context?: Manageme
 describe('oRPC Management capabilities route', () => {
 	test('reports every Management action available when the scheduler supports them', async () => {
 		const surface = createManagementSurface({
-			monque: createManagementMonque(),
+			monque: createManagementMonque({}, { mutations: true }),
 		});
 
 		const response = await handleCapabilities(surface);
@@ -78,7 +41,7 @@ describe('oRPC Management capabilities route', () => {
 	test('uses adapter-provided request context for action authorization', async () => {
 		const authorizedActions = new Set(['read', 'retry']);
 		const surface = createManagementSurface<{ role: string }>({
-			monque: createManagementMonque(),
+			monque: createManagementMonque({}, { mutations: true }),
 			authorize: ({ action, context }) => {
 				expect(context).toEqual({ role: 'viewer' });
 				return authorizedActions.has(action);
@@ -104,7 +67,7 @@ describe('oRPC Management capabilities route', () => {
 
 	test('reports writable actions unavailable in read-only mode', async () => {
 		const surface = createManagementSurface({
-			monque: createManagementMonque(),
+			monque: createManagementMonque({}, { mutations: true }),
 			readOnly: true,
 		});
 
@@ -125,13 +88,10 @@ describe('oRPC Management capabilities route', () => {
 
 	test('keeps unsupported actions visible as unavailable capabilities', async () => {
 		const surface = createManagementSurface({
-			monque: createManagementMonque(
-				{
-					retryJob: async () => null,
-					retryJobs: async () => ({ count: 0, errors: [] }),
-				},
-				{ mutations: false },
-			),
+			monque: createManagementMonque({
+				retryJob: async () => null,
+				retryJobs: async () => ({ count: 0, errors: [] }),
+			}),
 		});
 
 		const response = await handleCapabilities(surface);
