+++ packages/management/tests/unit/orpc-read-routes.test.ts
@@ -8,31 +8,9 @@ import {
 import { ObjectId } from 'mongodb';
 import { describe, expect, test } from 'vitest';
 
+import { createManagementMonque } from '@tests/unit/management-test-utils';
 import { createManagementSurface } from '@/index';
-import type { ManagementMonque, ManagementOpenApiContext, ManagementSurface } from '@/surface';
-
-function createManagementMonque(overrides: Partial<ManagementMonque> = {}): ManagementMonque {
-	return {
-		isHealthy: () => true,
-		getQueueViewSummaries: async () => [],
-		getJobsWithCursor: async () => ({
-			jobs: [],
-			cursor: null,
-			hasNextPage: false,
-			hasPreviousPage: false,
-		}),
-		getJob: async () => null,
-		getQueueStats: async () => ({
-			pending: 0,
-			processing: 0,
-			completed: 0,
-			failed: 0,
-			cancelled: 0,
-			total: 0,
-		}),
-		...overrides,
-	};
-}
+import type { ManagementOpenApiContext, ManagementSurface } from '@/surface';
 
 async function handleGet(
 	surface: ManagementSurface,
